+++ src/algokit/cli/task.py
@@ -2,6 +2,7 @@ import logging
 
 import click
 
+from algokit.cli.tasks.analyze import analyze
 from algokit.cli.tasks.assets import opt_in_command, opt_out_command
 from algokit.cli.tasks.ipfs import ipfs_group
 from algokit.cli.tasks.mint import mint
@@ -15,7 +16,19 @@ from algokit.cli.tasks.wallet import wallet
 logger = logging.getLogger(__name__)
 
 
-@click.group(name="task")
+class AliasedGroup(click.Group):
+    def get_command(self, ctx: click.Context, cmd_name: str) -> click.Command | None:
+        rv = click.Group.get_command(self, ctx, cmd_name)
+        if rv is not None:
+            return rv
+
+        if cmd_name == "analyse":
+            return click.Group.get_command(self, ctx, "analyze")
+
+        return None
+
+
+@click.group(name="task", cls=AliasedGroup)
 def task_group() -> None:
     """Collection of useful tasks to help you develop on Algorand."""
 
@@ -30,3 +43,4 @@ task_group.add_command(nfd_lookup)
 task_group.add_command(opt_out_command)
 task_group.add_command(opt_in_command)
 task_group.add_command(mint)
+task_group.add_command(analyze)